+++ lib/attributor/types/boolean.rb
@@ -16,6 +16,8 @@ module Attributor
     end
 
     def self.load(value,context=Attributor::DEFAULT_ROOT_CONTEXT, **options)
+      return nil if value.nil?
+
       raise CoercionError, context: context, from: value.class, to: self, value: value  if value.is_a?(::Float)
       return false if [ false, 'false', 'FALSE', '0', 0, 'f', 'F' ].include?(value)
       return true if [ true, 'true', 'TRUE', '1', 1, 't', 'T' ].include?(value)