+++ spec/types/boolean_spec.rb
@@ -64,6 +64,10 @@ describe Attributor::Boolean do
 
     end
 
+    it 'returns nil for nil' do
+      type.load(nil).should be(nil)
+    end
+
     context 'that are not valid Booleans' do
       let(:context){ ['root','subattr'] }
       ['string', 2, 1.0, Class, Object.new].each do |value|