+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Attributor Changelog
 next
 ------
 
+* Added `recurse` option to `Type.load` that is used by `Model` and `Hash` to force the loading of values (specifically, so that default values are assigned) even if the loaded value is `nil`.
 
 2.2.1
 ------